Crate hopli

hopli is a collection of commands to help with identity creation, funding, registration, etc. for HOPR nodes

Modules

environment_config

This module contains definiation of arguments that specify the environment and networks that a HOPR node runs in.

faucet

This module contains arguments and functions to fund some Ethereum wallets with native tokens and HOPR tokens

identity

This module contains subcommands for hopli identity, This command contains subcommands to read, create or update identity files, providing correct crate::key_pair::PasswordArgs

key_pair

This module contains struct definition, utility functions around private keys, password, and keystores.

methods

This module contains all the methods used for onchain interaction, especially with Safe instance, Mutlicall, and Multisend contracts.

network_registry

This module contains arguments and functions to interact with the Network Registry contract for a previledged account. To participate in the HOPR network, a node must be included in the network registry contract. Nodes and the staking account (Safe) that manages them should be registered as a pair in the Network registry contrat. Nodes and safes can be registered by either a manager or by the staking account itself.

safe_module

This module contains arguments and functions to manage safe and module. SafeModuleSubcommands defines three subcommands: create, move, and migrate.

utils

This module contains errors produced in this crate

Structs

Cli 🔒

Enums

Commands 🔒

Helper for running your HOPR nodes

Functions

main 🔒