hopli

Module constants

Constants

DEFAULT_ANNOUNCEMENT_PERMISSIONS

Default announcement permissions, where nothing is specified and falls back to the default, as in smart contract

DEFAULT_CAPABILITY_PERMISSIONS

Default capability permissions, as in smart contract

DEFAULT_NODE_PERMISSIONS

Default node permissions, where nothing is specified and falls back to the default, as in smart contract

DIV_BY_ZERO

Panic(0x12) division by zero https://docs.soliditylang.org/en/v0.8.12/control-structures.html#panic-via-assert-and-error-via-require

DOMAIN_SEPARATOR_TYPEHASH

Safe domain separator typehash, as in https://github.com/safe-global/safe-smart-account/blob/2278f7ccd502878feb5cec21dd6255b82df374b5/contracts/Safe.sol#L54

ERC_1820_DEPLOYER

ERC1820 deployer wallet

ERC_1820_REGISTRY_DEPLOY_CODE

Deployed code of the ERC1820 registry contract of type [u8; 2619]

ERC_1820_SIGNED_DEPLOYMENT_TX

Signed tx to create an ERC1820 contract

ERC_1967_PROXY_CREATION_CODE

ETH_VALUE_FOR_ERC1820_DEPLOYER

Amount of ETH to fund ERC_1820_DEPLOYER, to deployer ERC1820 registry ERC1820Registry contract is used by the HOPR token contract (ERC777 token)

ETH_VALUE_FOR_MULTICALL3_DEPLOYER

INIT_KEY_BINDING_FEE

Initial key binding fee required to initialize a key binding in the HOPR token contract

MULTICALL3_DEPLOYER

MULTICALL3_DEPLOY_CODE

SAFE_COMPATIBILITYFALLBACKHANDLER_ADDRESS

Safe compatibility fallback handler contract deployed for v1.5.0, as in https://github.com/safe-global/safe-deployments/tree/7a868b24393860b1bc5db8811f4cbbc10c58032d/src/assets/v1.5.0

SAFE_COMPATIBILITY_FALLBACK_HANDLER_DEPLOY_CODE

SAFE_DIAMOND_PROXY_SINGLETON_DEPLOY_CODE

Deployed code of the Safe Diamond Proxy Singleton

SAFE_MULTISEND_ADDRESS

Safe multisend contract deployed for v1.4.0, https://github.com/safe-global/safe-deployments/blob/b707e5e2994e6f86d76ff7ffade0445c4e49ae9a/src/assets/v1.4.0/multi_send.json

SAFE_MULTISEND_CALL_ONLY_DEPLOY_CODE

SAFE_PROXY_FACTORY_DEPLOY_CODE

SAFE_SAFEPROXYFACTORY_ADDRESS

Safe proxy factory contract deployed for v1.4.0, https://github.com/safe-global/safe-deployments/blob/b707e5e2994e6f86d76ff7ffade0445c4e49ae9a/src/assets/v1.4.0/safe_proxy_factory.json

SAFE_SAFE_L2_ADDRESS

Safe implementation contract (L2) deployed for v1.4.1, as in https://github.com/safe-global/safe-deployments/blob/8c504f44d148f4c898fee02749c88372bae6609a/src/assets/v1.4.1/safe.json

SAFE_TX_TYPEHASH

Safe transaction type hash, as in https://github.com/safe-global/safe-smart-account/blob/2278f7ccd502878feb5cec21dd6255b82df374b5/contracts/Safe.sol#L59

SENTINEL_OWNERS

Starting point for a linked list in safe contract, as in https://github.com/safe-global/safe-smart-account/blob/2278f7ccd502878feb5cec21dd6255b82df374b5/contracts/base/OwnerManager.sol#L14

Statics

SAFE_SINGLETON_DEPLOY_CODE