hopli

Module identity

This module contains subcommands for hopli identity, This command contains subcommands to read, create or update identity files, providing correct crate::key_pair::PasswordArgs

For all three actions around identity files, at least two arguments are needed:

Path to files

Password to encrypt or decrypt the file

To update identity files, a new password must be provided.

Some sample commands:

To create identities

hopli identity create \
    --identity-directory "./test" \
    --identity-prefix nodes_ \
    --number 2 \
    --password-path "./test/pwd"

To read identities

hopli identity read \
    --identity-directory "./test" \
    --identity-prefix node_ \
    --password-path "./test/pwd"

To update password of identities

    hopli identity update \
    --identity-directory "./test" \
    --identity-prefix node_ \
    --password-path "./test/pwd" \
    --new-password-path "./test/newpwd"

Structs

ConvertPeerArgs

Arguments for PeerID or Public key

Enums

IdentitySubcommands

CLI arguments for hopli identity