hopli::methods

Module ModuleSingleton

Module containing a contract’s types and functions.

contract ModuleSingleton {
    function isNode(address) external view returns (bool);
    function getTargets() external view returns (uint256[] memory);
    function owner() public view returns (address);
}

Modules

abi

Contains dynamic ABI definitions for this contract.

Structs

ModuleSingletonInstance

A ModuleSingleton instance.

getTargetsCall

Function with signature getTargets() and selector 0x63fe3b56.

getTargetsReturn

Container type for the return parameters of the getTargets() function.

isNodeCall

Function with signature isNode(address) and selector 0x01750152.

isNodeReturn

Container type for the return parameters of the isNode(address) function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

Enums

ModuleSingletonCalls

Container for all the ModuleSingleton function calls.

Functions

new

Creates a new wrapper around an on-chain ModuleSingleton contract instance.