hopli::methods

Module SafeSingleton

Module containing a contract’s types and functions.

contract SafeSingleton {
    event ExecutionSuccess(bytes32 indexed txHash, uint256 payment);
    function setup(address[], uint256, address, bytes, address, address, uint256, address);
    function execTransaction(address to, uint256 value, bytes calldata data, uint8 operation, uint256 safeTxGas, uint256 baseGas, uint256 gasPrice, address gasToken, address refundReceiver, bytes memory signatures) public payable returns (bool);
    function removeOwner(address prevOwner, address owner, uint256 _threshold) public;
    function getThreshold() public view returns (uint256);
    function getOwners() public view returns (address[] memory);
    function nonce() public view returns (uint256);
    function domainSeparator() public view returns (bytes32);
    function encodeTransactionData(address to, uint256 value, bytes calldata data, uint8 operation, uint256 safeTxGas, uint256 baseGas, uint256 gasPrice, address gasToken, address refundReceiver, uint256 _nonce) public view returns (bytes memory);
    function getTransactionHash(address to, uint256 value, bytes calldata data, uint8 operation, uint256 safeTxGas, uint256 baseGas, uint256 gasPrice, address gasToken, address refundReceiver, uint256 _nonce) public view returns (bytes32);
    function isModuleEnabled(address module) public view returns (bool);
}

Modules

abi

Contains dynamic ABI definitions for this contract.

Structs

ExecutionSuccess

Event with signature ExecutionSuccess(bytes32,uint256) and selector 0x442e715f626346e8c54381002da614f62bee8d27386535b2521ec8540898556e.

SafeSingletonInstance

A SafeSingleton instance.

domainSeparatorCall

Function with signature domainSeparator() and selector 0xf698da25.

domainSeparatorReturn

Container type for the return parameters of the domainSeparator() function.

encodeTransactionDataCall

Function with signature encodeTransactionData(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,uint256) and selector 0xe86637db.

encodeTransactionDataReturn

Container type for the return parameters of the encodeTransactionData(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,uint256) function.

execTransactionCall

Function with signature execTransaction(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,bytes) and selector 0x6a761202.

execTransactionReturn

Container type for the return parameters of the execTransaction(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,bytes) function.

getOwnersCall

Function with signature getOwners() and selector 0xa0e67e2b.

getOwnersReturn

Container type for the return parameters of the getOwners() function.

getThresholdCall

Function with signature getThreshold() and selector 0xe75235b8.

getThresholdReturn

Container type for the return parameters of the getThreshold() function.

getTransactionHashCall

Function with signature getTransactionHash(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,uint256) and selector 0xd8d11f78.

getTransactionHashReturn

Container type for the return parameters of the getTransactionHash(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,uint256) function.

isModuleEnabledCall

Function with signature isModuleEnabled(address) and selector 0x2d9ad53d.

isModuleEnabledReturn

Container type for the return parameters of the isModuleEnabled(address) function.

nonceCall

Function with signature nonce() and selector 0xaffed0e0.

nonceReturn

Container type for the return parameters of the nonce() function.

removeOwnerCall

Function with signature removeOwner(address,address,uint256) and selector 0xf8dc5dd9.

removeOwnerReturn

Container type for the return parameters of the removeOwner(address,address,uint256) function.

setupCall

Function with signature setup(address[],uint256,address,bytes,address,address,uint256,address) and selector 0xb63e800d.

setupReturn

Container type for the return parameters of the setup(address[],uint256,address,bytes,address,address,uint256,address) function.

Enums

SafeSingletonCalls

Container for all the SafeSingleton function calls.

SafeSingletonEvents

Container for all the SafeSingleton events.

Functions

new

Creates a new wrapper around an on-chain SafeSingleton contract instance.