hopli::methods

Function deregister_nodes_from_node_safe_registry_and_remove_from_module

pub async fn deregister_nodes_from_node_safe_registry_and_remove_from_module<M: Middleware>(
    node_safe_registry: HoprNodeSafeRegistry<M>,
    node_addresses: Vec<H160>,
    module_addresses: Vec<H160>,
    owner_chain_key: ChainKeypair,
) -> Result<u32, HelperErrors>

Deregister safes and nodes from the node-safe registry. It returns the number of removed nodes

If nodes have been registered to a safe, remove the node

If nodes have not been registered to any safe, no op

When deregsitering one node, also remove the node from the module