hopli::methods

Function register_safes_and_nodes_on_network_registry

pub async fn register_safes_and_nodes_on_network_registry<P: Provider + WalletProvider + Clone>(
    network_registry: HoprNetworkRegistryInstance<Arc<P>>,
    safe_addresses: Vec<Address>,
    node_addresses: Vec<Address>,
) -> Result<(usize, usize), HelperErrors>

Register safes and nodes to the network registry, and force-sync the eligibility to true. It returns the number of removed nodes and nodes being added.

If nodes have been registered to a different safe, overwrite it (remove the old safe and regsiter with the new safe)

If ndoes have been registered to the same safe, no op

If nodes have not been registered to any safe, register it

After all the nodes have been added to the network registry, force-sync the eligibility of all the added safes to true