hopli::methods

Function send_multisend_safe_transaction_with_threshold_one

pub async fn send_multisend_safe_transaction_with_threshold_one<P: WalletProvider + Provider>(
    safe: SafeSingletonInstance<Arc<P>>,
    signer_key: ChainKeypair,
    multisend_contract: Address,
    multisend_txns: Vec<MultisendTransaction>,
    chain_id: U256,
    nonce: U256,
) -> Result<(), HelperErrors>

Use safe to delegatecall to multisend contract Note that when no additional signature is provided, the safe must have a threshold of one, so that the transaction can be executed. Note that the refund address is the caller (safe owner) wallet