hopli::methods
Module module_singleton

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

GetTargetsCall

Container type for all input parameters for the getTargets function with signature getTargets() and selector 0x63fe3b56

GetTargetsReturn

Container type for all return fields from the getTargets function with signature getTargets() and selector 0x63fe3b56

IsNodeCall

Container type for all input parameters for the isNode function with signature isNode(address) and selector 0x01750152

IsNodeReturn

Container type for all return fields from the isNode function with signature isNode(address) and selector 0x01750152

ModuleSingleton

OwnerCall

Container type for all input parameters for the owner function with signature owner() and selector 0x8da5cb5b

OwnerReturn

Container type for all return fields from the owner function with signature owner() and selector 0x8da5cb5b

Enums

ModuleSingletonCalls

Container type for all of the contract’s call

Statics

MODULESINGLETON_ABI

The parsed human-readable ABI of the contract.

Functions

__abi 🔒