hopli::methods

Module safe_singleton

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

DomainSeparatorCall

Container type for all input parameters for the domainSeparator function with signature domainSeparator() and selector 0xf698da25

DomainSeparatorReturn

Container type for all return fields from the domainSeparator function with signature domainSeparator() and selector 0xf698da25

EncodeTransactionDataCall

Container type for all input parameters for the encodeTransactionData function with signature encodeTransactionData(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,uint256) and selector 0xe86637db

EncodeTransactionDataReturn

Container type for all return fields from the encodeTransactionData function with signature encodeTransactionData(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,uint256) and selector 0xe86637db

ExecTransactionCall

Container type for all input parameters for the execTransaction function with signature execTransaction(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,bytes) and selector 0x6a761202

ExecTransactionReturn

Container type for all return fields from the execTransaction function with signature execTransaction(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,bytes) and selector 0x6a761202

GetOwnersCall

Container type for all input parameters for the getOwners function with signature getOwners() and selector 0xa0e67e2b

GetOwnersReturn

Container type for all return fields from the getOwners function with signature getOwners() and selector 0xa0e67e2b

GetThresholdCall

Container type for all input parameters for the getThreshold function with signature getThreshold() and selector 0xe75235b8

GetThresholdReturn

Container type for all return fields from the getThreshold function with signature getThreshold() and selector 0xe75235b8

GetTransactionHashCall

Container type for all input parameters for the getTransactionHash function with signature getTransactionHash(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,uint256) and selector 0xd8d11f78

GetTransactionHashReturn

Container type for all return fields from the getTransactionHash function with signature getTransactionHash(address,uint256,bytes,uint8,uint256,uint256,uint256,address,address,uint256) and selector 0xd8d11f78

IsModuleEnabledCall

Container type for all input parameters for the isModuleEnabled function with signature isModuleEnabled(address) and selector 0x2d9ad53d

IsModuleEnabledReturn

Container type for all return fields from the isModuleEnabled function with signature isModuleEnabled(address) and selector 0x2d9ad53d

NonceCall

Container type for all input parameters for the nonce function with signature nonce() and selector 0xaffed0e0

NonceReturn

Container type for all return fields from the nonce function with signature nonce() and selector 0xaffed0e0

RemoveOwnerCall

Container type for all input parameters for the removeOwner function with signature removeOwner(address,address,uint256) and selector 0xf8dc5dd9

SafeSingleton

SetupCall

Container type for all input parameters for the setup function with signature setup(address[],uint256,address,bytes,address,address,uint256,address) and selector 0xb63e800d

Enums

SafeSingletonCalls

Container type for all of the contract’s call

Statics

SAFESINGLETON_ABI

The parsed human-readable ABI of the contract.

Functions

__abi 🔒