hopli

Module environment_config

This module contains definiation of arguments that specify the environment and networks that a HOPR node runs in.

EnvironmentType defines the environment type. EnvironmentType of a network is defined in contracts-addresses.json under the foundry contract root. Different environment type uses a different foundry profile.

Network is a collection of several major/minor releases.

NetworkDetail specifies the environment type of the network, the starting block number, and the deployed contract addresses in ContractAddresses

Structs

NetworkConfig

mapping of networks with its details

NetworkDetail

NetworkProviderArgs

Arguments for getting network and ethereum RPC provider.

Functions

ensure_environment_and_network_are_set

ensures that the network and environment_type exist in contracts-addresses.json and are matched

get_environment_type_from_name

Returns the environment type from the network name according to contracts-addresses.json

get_network_details_from_name

Get the NetworkDetail (contract addresses, environment type) from network names

Type Aliases

RpcProvider

RpcProviderWithoutSigner

SharedFillerChain 🔒