hopli

Module key_pair

This module contains struct definition, utility functions around private keys, password, and keystores.

Keystore file is often referred as HOPR node identity file, which is an encrypted private key for an Ethereum wallet. This identity file uses password (received from PasswordArgs) for encryption.

Location of identity files can be provided with IdentityFileArgs.

This module also contains definition of argument for private key, defined in PrivateKeyArgs.

Structs

IdentityFileArgs

CLI arguments to specify the directory of one or multiple identity files

IdentityFromDirectoryArgs

CLI arguments to specify the directory of one or multiple identity files

ManagerPrivateKeyArgs

Arguments for private key.

NewPasswordArgs

Arguments for new password.

PasswordArgs

Arguments for password.

PrivateKeyArgs

Arguments for private key.

Traits

ArgEnvReader

Functions

create_identity

Create one identity file and return the ethereum address

read_identities

Decrypts identity files and returns an vec of PeerIds and Ethereum Addresses

read_identity

update_identity_password

encrypt HoprKeys with a new password to an identity file