hopli
Module utils

This module contains errors produced in this crate

Enums

HelperErrors

Enumerates different errors produced by this crate.

Constants

DEFAULT_ANNOUNCEMENT_PERMISSIONS

Default announcement permissions, where nothing is specified and falls back to the default, as in smart contract

DEFAULT_CAPABILITY_PERMISSIONS

Default capability permissions, as in smart contract

DEFAULT_NODE_PERMISSIONS

Default node permissions, where nothing is specified and falls back to the default, as in smart contract

DOMAIN_SEPARATOR_TYPEHASH

Safe domain separator typehash, as in https://github.com/safe-global/safe-smart-account/blob/2278f7ccd502878feb5cec21dd6255b82df374b5/contracts/Safe.sol#L54

ETH_VALUE_FOR_MULTICALL3_DEPLOYER

Amount of ETH to fund MULTICALL3_DEPLOYER, to deployer Multicall3 contract

MULTICALL3_CONTRACT_CREATION_CODE

Contract creation code for mullticall3, as in https://etherscan.io/address/0xcA11bde05977b3631167028862bE2a173976CA11

MULTICALL3_DEPLOYER

Multicall3 deployer wallet

NEW_HOPR_NODE_STAKE_MODULE_TOPIC

Topic hash for NewHoprNodeStakeModule event

NEW_HOPR_NODE_STAKE_SAFE_TOPIC

Topic hash for NewHoprNodeStakeSafe event

SAFE_COMPATIBILITYFALLBACKHANDLER_ADDRESS

Safe compatibility fallback handler contract deployed for v1.4.0, as in https://github.com/safe-global/safe-deployments/blob/b707e5e2994e6f86d76ff7ffade0445c4e49ae9a/src/assets/v1.4.0/compatibility_fallback_handler.json

SAFE_EXECUTION_SUCCESS

Topic hash for ExecutionSuccess event, as in https://github.com/safe-global/safe-smart-account/blob/2278f7ccd502878feb5cec21dd6255b82df374b5/contracts/interfaces/ISafe.sol#L18

SAFE_MULTISEND_ADDRESS

Safe multisend contract deployed for v1.4.0, https://github.com/safe-global/safe-deployments/blob/b707e5e2994e6f86d76ff7ffade0445c4e49ae9a/src/assets/v1.4.0/multi_send.json

SAFE_SAFEPROXYFACTORY_ADDRESS

Safe proxy factory contract deployed for v1.4.0, https://github.com/safe-global/safe-deployments/blob/b707e5e2994e6f86d76ff7ffade0445c4e49ae9a/src/assets/v1.4.0/safe_proxy_factory.json

SAFE_SAFE_ADDRESS

Safe implementation contract deployed for v1.4.1, as in https://github.com/safe-global/safe-deployments/blob/8c504f44d148f4c898fee02749c88372bae6609a/src/assets/v1.4.1/safe.json

SAFE_TX_TYPEHASH

Safe transaction type hash, as in https://github.com/safe-global/safe-smart-account/blob/2278f7ccd502878feb5cec21dd6255b82df374b5/contracts/Safe.sol#L59

SENTINEL_OWNERS

Starting point for a linked list in safe contract, as in https://github.com/safe-global/safe-smart-account/blob/2278f7ccd502878feb5cec21dd6255b82df374b5/contracts/base/OwnerManager.sol#L14

Traits

Cmd